* src/m4.h (hash_table_size): Now size_t instead of int. * src/m4.c (hash_table_size): Likewise. (main): Adjust to this; use atol rather than atoi. * src/symtab.c: Include <limits.h>, for CHAR_BIT. (symtab_init, lookup_symbol, hack_all_symbols): Use size_t for sizes and indexes, not int. (symtab_print_list) [DEBUG_SYM]: Likewise. (hash): Likewise.  Don't case-fold in the hash function. Shift by 7, not 3, for consistency with gnulib/lib/hash.c. Don't assume hash word is 32 bits. NEWS: Document this change.

+++ src/m4.h
@@ -147,7 +147,7 @@ typedef struct token_data token_data;
 /* Option flags.  */
 extern int sync_output;			/* -s */
 extern int debug_level;			/* -d */
-extern int hash_table_size;		/* -H */
+extern size_t hash_table_size;		/* -H */
 extern int no_gnu_extensions;		/* -G */
 extern int prefix_all_builtins;		/* -P */
 extern int max_debug_argument_length;	/* -l */
